Alterações dia 15/05/2019
