+++ nbproject/project.properties
@@ -28,12 +28,14 @@ dist.jar=${dist.dir}/Gabenricks_H_C.jar
 dist.javadoc.dir=${dist.dir}/javadoc
 excludes=
 file.reference.DateChooser.jar=C:\\Users\\tavar\\Downloads\\jdatechooser\\DateChooser.jar
+file.reference.DateChooser.jar-1=jdatechooser/DateChooser.jar
 file.reference.jcalendar-1.4.jar=C:\\Users\\tavar\\Downloads\\jcalendar-1.4.jar
 includes=**
 jar.compress=false
 javac.classpath=\
     ${file.reference.DateChooser.jar}:\
-    ${file.reference.jcalendar-1.4.jar}
+    ${file.reference.jcalendar-1.4.jar}:\
+    ${file.reference.DateChooser.jar-1}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
@@ -58,7 +60,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=gabenricks_h.c.Gabenricks_HC
+main.class=br.com.gabenricks_HeC.Views.Principal
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
