+++ src/br/com/gabenricks_HeC/Views/Cliente_Cadastro.java
@@ -1,0 +1,461 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package br.com.gabenricks_HeC.Views;
+
+import java.awt.Dimension;
+
+/**
+ *
+ * @author marcus.hmoraes
+ */
+public class Cliente_Cadastro extends javax.swing.JInternalFrame {
+
+    /**
+     * Creates new form Cliente_Cadastro
+     */
+    public Cliente_Cadastro() {
+        
+        initComponents();
+        Dimension size = dateChooserCombo1.getCalendarPreferredSize();
+        size.width += 90;
+        dateChooserCombo1.setCalendarPreferredSize(size);
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        DesktopMain = new javax.swing.JDesktopPane();
+        lblNome = new javax.swing.JLabel();
+        btnNomeCli = new javax.swing.JTextField();
+        lblDataNacimento = new javax.swing.JLabel();
+        lblCPF = new javax.swing.JLabel();
+        lblFone = new javax.swing.JLabel();
+        lblEmail = new javax.swing.JLabel();
+        lblCodFidelidade = new javax.swing.JLabel();
+        txtCPF = new javax.swing.JFormattedTextField();
+        txtFone = new javax.swing.JFormattedTextField();
+        txtEmail = new javax.swing.JTextField();
+        txtCodFidelidade = new javax.swing.JTextField();
+        lblEndereco = new javax.swing.JLabel();
+        lblNumero = new javax.swing.JLabel();
+        lblCEP = new javax.swing.JLabel();
+        lblComplemento = new javax.swing.JLabel();
+        lblCidade = new javax.swing.JLabel();
+        lblEstado = new javax.swing.JLabel();
+        CboEndereco = new javax.swing.JComboBox<>();
+        txtEndereco = new javax.swing.JTextField();
+        txtComplemento = new javax.swing.JTextField();
+        txtCidade = new javax.swing.JTextField();
+        jComboBox2 = new javax.swing.JComboBox<>();
+        AtxtComplemento = new javax.swing.JTextArea();
+        lblInfComplement = new javax.swing.JLabel();
+        txtCEP = new javax.swing.JFormattedTextField();
+        jFormattedTextField1 = new javax.swing.JFormattedTextField();
+        dateChooserCombo1 = new datechooser.beans.DateChooserCombo();
+        jButton1 = new javax.swing.JButton();
+        jButton2 = new javax.swing.JButton();
+
+        setBorder(null);
+        setDefaultCloseOperation(javax.swing.WindowConstants.DO_NOTHING_ON_CLOSE);
+        setForeground(java.awt.Color.white);
+        setMaximizable(true);
+        setTitle("Cadastro de Cliente");
+        setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
+        setEnabled(false);
+        setFrameIcon(null);
+        setMaximumSize(new java.awt.Dimension(852, 852));
+        setMinimumSize(new java.awt.Dimension(852, 852));
+        try {
+            setSelected(true);
+        } catch (java.beans.PropertyVetoException e1) {
+            e1.printStackTrace();
+        }
+        setVerifyInputWhenFocusTarget(false);
+        setVisible(true);
+
+        DesktopMain.setBackground(new java.awt.Color(240, 240, 240));
+        DesktopMain.setForeground(new java.awt.Color(240, 240, 240));
+        DesktopMain.setEnabled(false);
+        DesktopMain.setPreferredSize(new java.awt.Dimension(846, 500));
+
+        lblNome.setText("Nome");
+
+        btnNomeCli.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnNomeCliActionPerformed(evt);
+            }
+        });
+
+        lblDataNacimento.setText("Data de nascimento");
+
+        lblCPF.setText("CPF");
+
+        lblFone.setText("Telefone");
+
+        lblEmail.setText("E-mail");
+
+        lblCodFidelidade.setText("Código de fidelidade");
+
+        try {
+            txtCPF.setFormatterFactory(new javax.swing.text.DefaultFormatterFactory(new javax.swing.text.MaskFormatter("###.###.###-##")));
+        } catch (java.text.ParseException ex) {
+            ex.printStackTrace();
+        }
+        txtCPF.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                txtCPFActionPerformed(evt);
+            }
+        });
+
+        try {
+            txtFone.setFormatterFactory(new javax.swing.text.DefaultFormatterFactory(new javax.swing.text.MaskFormatter("(###)####-####")));
+        } catch (java.text.ParseException ex) {
+            ex.printStackTrace();
+        }
+
+        txtEmail.setToolTipText("Ex.: Senac@senac.edu.br");
+        txtEmail.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                txtEmailActionPerformed(evt);
+            }
+        });
+
+        txtCodFidelidade.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                txtCodFidelidadeActionPerformed(evt);
+            }
+        });
+
+        lblEndereco.setText("Endereço");
+
+        lblNumero.setText("Número");
+
+        lblCEP.setText("CEP");
+
+        lblComplemento.setText("Complemento");
+
+        lblCidade.setText("Cidade");
+
+        lblEstado.setText("Estado");
+
+        CboEndereco.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Rua", "Avenida" }));
+
+        txtEndereco.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                txtEnderecoActionPerformed(evt);
+            }
+        });
+
+        txtComplemento.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                txtComplementoActionPerformed(evt);
+            }
+        });
+
+        txtCidade.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                txtCidadeActionPerformed(evt);
+            }
+        });
+
+        jComboBox2.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "AC", "AL", "AM", "AP", "BA", "CE", "DF", "ES", "GO", "MA", "MT", "MS", "MG", "PA", "PB", "PR", "PE", "PI", "RJ", "RN", "RS", "RO", "RR", "SC", "SP", "SE", "TO" }));
+        jComboBox2.setToolTipText("Selecione seu Estado.");
+        jComboBox2.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jComboBox2ActionPerformed(evt);
+            }
+        });
+
+        AtxtComplemento.setColumns(20);
+        AtxtComplemento.setRows(5);
+
+        lblInfComplement.setText("Informações complementares");
+
+        try {
+            txtCEP.setFormatterFactory(new javax.swing.text.DefaultFormatterFactory(new javax.swing.text.MaskFormatter("######-###")));
+        } catch (java.text.ParseException ex) {
+            ex.printStackTrace();
+        }
+
+        jFormattedTextField1.setFormatterFactory(new javax.swing.text.DefaultFormatterFactory(new javax.swing.text.NumberFormatter(new java.text.DecimalFormat("#0"))));
+
+        dateChooserCombo1.setWeekStyle(datechooser.view.WeekDaysStyle.SHORT);
+        try {
+            dateChooserCombo1.setDefaultPeriods(new datechooser.model.multiple.PeriodSet(new datechooser.model.multiple.Period(new java.util.GregorianCalendar(2019, 4, 15),
+                new java.util.GregorianCalendar(2019, 4, 15))));
+    } catch (datechooser.model.exeptions.IncompatibleDataExeption e1) {
+        e1.printStackTrace();
+    }
+    dateChooserCombo1.setNavigateFont(new java.awt.Font("Serif", java.awt.Font.PLAIN, 7));
+    dateChooserCombo1.setBehavior(datechooser.model.multiple.MultyModelBehavior.SELECT_PERIOD);
+    dateChooserCombo1.addCommitListener(new datechooser.events.CommitListener() {
+        public void onCommit(datechooser.events.CommitEvent evt) {
+            dateChooserCombo1OnCommit(evt);
+        }
+    });
+
+    jButton1.setText("jButton1");
+
+    jButton2.setText("jButton2");
+
+    DesktopMain.setLayer(lblNome, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(btnNomeCli, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(lblDataNacimento, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(lblCPF, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(lblFone, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(lblEmail, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(lblCodFidelidade, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(txtCPF, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(txtFone, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(txtEmail, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(txtCodFidelidade, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(lblEndereco, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(lblNumero, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(lblCEP, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(lblComplemento, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(lblCidade, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(lblEstado, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(CboEndereco, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(txtEndereco, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(txtComplemento, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(txtCidade, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(jComboBox2, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(AtxtComplemento, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(lblInfComplement, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(txtCEP, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(jFormattedTextField1, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(dateChooserCombo1, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(jButton1, javax.swing.JLayeredPane.DEFAULT_LAYER);
+    DesktopMain.setLayer(jButton2, javax.swing.JLayeredPane.DEFAULT_LAYER);
+
+    javax.swing.GroupLayout DesktopMainLayout = new javax.swing.GroupLayout(DesktopMain);
+    DesktopMain.setLayout(DesktopMainLayout);
+    DesktopMainLayout.setHorizontalGroup(
+        DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+        .addGroup(DesktopMainLayout.createSequentialGroup()
+            .addGap(30, 30, 30)
+            .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(DesktopMainLayout.createSequentialGroup()
+                    .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                        .addComponent(btnNomeCli)
+                        .addGroup(DesktopMainLayout.createSequentialGroup()
+                            .addComponent(lblNome)
+                            .addGap(0, 0, Short.MAX_VALUE))
+                        .addGroup(DesktopMainLayout.createSequentialGroup()
+                            .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                .addGroup(DesktopMainLayout.createSequentialGroup()
+                                    .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                        .addComponent(lblEndereco)
+                                        .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
+                                            .addGroup(DesktopMainLayout.createSequentialGroup()
+                                                .addComponent(CboEndereco, javax.swing.GroupLayout.PREFERRED_SIZE, 69, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                                                .addComponent(txtEndereco))
+                                            .addComponent(txtEmail, javax.swing.GroupLayout.Alignment.LEADING)
+                                            .addGroup(javax.swing.GroupLayout.Alignment.LEADING, DesktopMainLayout.createSequentialGroup()
+                                                .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                                    .addComponent(lblEmail)
+                                                    .addComponent(lblDataNacimento, javax.swing.GroupLayout.DEFAULT_SIZE, 108, Short.MAX_VALUE)
+                                                    .addComponent(dateChooserCombo1, javax.swing.GroupLayout.PREFERRED_SIZE, 0, Short.MAX_VALUE))
+                                                .addGap(53, 53, 53)
+                                                .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                                    .addComponent(txtCPF, javax.swing.GroupLayout.DEFAULT_SIZE, 324, Short.MAX_VALUE)
+                                                    .addComponent(lblCPF, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)))))
+                                    .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 57, Short.MAX_VALUE))
+                                .addGroup(DesktopMainLayout.createSequentialGroup()
+                                    .addComponent(lblComplemento)
+                                    .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                    .addComponent(lblCidade)
+                                    .addGap(75, 75, 75)))
+                            .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                .addComponent(txtFone)
+                                .addComponent(lblCodFidelidade)
+                                .addComponent(txtCodFidelidade)
+                                .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, DesktopMainLayout.createSequentialGroup()
+                                    .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                                        .addComponent(lblNumero, javax.swing.GroupLayout.PREFERRED_SIZE, 82, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                        .addComponent(jFormattedTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, 82, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                    .addGap(18, 18, 18)
+                                    .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                        .addGroup(DesktopMainLayout.createSequentialGroup()
+                                            .addComponent(lblCEP)
+                                            .addGap(103, 103, 103))
+                                        .addComponent(txtCEP)))
+                                .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                    .addComponent(lblEstado)
+                                    .addComponent(jComboBox2, javax.swing.GroupLayout.PREFERRED_SIZE, 69, javax.swing.GroupLayout.PREFERRED_SIZE))
+                                .addComponent(lblFone, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))))
+                    .addGap(88, 88, 88))
+                .addGroup(DesktopMainLayout.createSequentialGroup()
+                    .addComponent(lblInfComplement)
+                    .addGap(711, 711, Short.MAX_VALUE))
+                .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, DesktopMainLayout.createSequentialGroup()
+                    .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
+                        .addGroup(DesktopMainLayout.createSequentialGroup()
+                            .addComponent(AtxtComplemento, javax.swing.GroupLayout.PREFERRED_SIZE, 385, javax.swing.GroupLayout.PREFERRED_SIZE)
+                            .addGap(0, 298, Short.MAX_VALUE))
+                        .addGroup(DesktopMainLayout.createSequentialGroup()
+                            .addComponent(txtComplemento)
+                            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                            .addComponent(txtCidade, javax.swing.GroupLayout.PREFERRED_SIZE, 254, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                    .addGap(169, 169, 169))))
+        .addGroup(DesktopMainLayout.createSequentialGroup()
+            .addGap(55, 55, 55)
+            .addComponent(jButton1)
+            .addGap(44, 44, 44)
+            .addComponent(jButton2)
+            .addGap(0, 0, Short.MAX_VALUE))
+    );
+    DesktopMainLayout.setVerticalGroup(
+        DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+        .addGroup(DesktopMainLayout.createSequentialGroup()
+            .addContainerGap()
+            .addComponent(lblNome)
+            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+            .addComponent(btnNomeCli, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+            .addGap(18, 18, 18)
+            .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                .addComponent(lblDataNacimento)
+                .addComponent(lblCPF)
+                .addComponent(lblFone))
+            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+            .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(txtCPF, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(txtFone, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addComponent(dateChooserCombo1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+            .addGap(18, 18, 18)
+            .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                .addComponent(lblEmail, javax.swing.GroupLayout.PREFERRED_SIZE, 14, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(lblCodFidelidade, javax.swing.GroupLayout.PREFERRED_SIZE, 14, javax.swing.GroupLayout.PREFERRED_SIZE))
+            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+            .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                .addComponent(txtEmail, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(txtCodFidelidade, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+            .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addComponent(lblNumero, javax.swing.GroupLayout.PREFERRED_SIZE, 14, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(lblEndereco, javax.swing.GroupLayout.PREFERRED_SIZE, 14, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(lblCEP, javax.swing.GroupLayout.PREFERRED_SIZE, 14, javax.swing.GroupLayout.PREFERRED_SIZE)))
+            .addGap(10, 10, 10)
+            .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                .addComponent(CboEndereco, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(txtEndereco, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(txtCEP, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(jFormattedTextField1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+            .addGap(18, 18, 18)
+            .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                .addComponent(lblComplemento, javax.swing.GroupLayout.PREFERRED_SIZE, 14, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(lblCidade, javax.swing.GroupLayout.PREFERRED_SIZE, 14, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(lblEstado, javax.swing.GroupLayout.PREFERRED_SIZE, 14, javax.swing.GroupLayout.PREFERRED_SIZE))
+            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+            .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                .addComponent(txtComplemento, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(txtCidade, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addComponent(jComboBox2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+            .addGap(18, 18, 18)
+            .addComponent(lblInfComplement, javax.swing.GroupLayout.PREFERRED_SIZE, 14, javax.swing.GroupLayout.PREFERRED_SIZE)
+            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+            .addComponent(AtxtComplemento, javax.swing.GroupLayout.PREFERRED_SIZE, 124, javax.swing.GroupLayout.PREFERRED_SIZE)
+            .addGap(18, 18, 18)
+            .addGroup(DesktopMainLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addComponent(jButton1)
+                .addComponent(jButton2))
+            .addContainerGap(32, Short.MAX_VALUE))
+    );
+
+    javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
+    getContentPane().setLayout(layout);
+    layout.setHorizontalGroup(
+        layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+        .addComponent(DesktopMain, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, 882, Short.MAX_VALUE)
+    );
+    layout.setVerticalGroup(
+        layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+        .addComponent(DesktopMain, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, 525, Short.MAX_VALUE)
+    );
+
+    setBounds(0, 0, 882, 552);
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void btnNomeCliActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnNomeCliActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_btnNomeCliActionPerformed
+
+    private void txtEmailActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txtEmailActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_txtEmailActionPerformed
+
+    private void txtCodFidelidadeActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txtCodFidelidadeActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_txtCodFidelidadeActionPerformed
+
+    private void txtEnderecoActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txtEnderecoActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_txtEnderecoActionPerformed
+
+    private void txtComplementoActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txtComplementoActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_txtComplementoActionPerformed
+
+    private void txtCidadeActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txtCidadeActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_txtCidadeActionPerformed
+
+    private void txtCPFActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_txtCPFActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_txtCPFActionPerformed
+
+
+    private void jComboBox2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jComboBox2ActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_jComboBox2ActionPerformed
+
+    private void dateChooserCombo1OnCommit(datechooser.events.CommitEvent evt) {//GEN-FIRST:event_dateChooserCombo1OnCommit
+        // TODO add your handling code here:
+    }//GEN-LAST:event_dateChooserCombo1OnCommit
+
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JTextArea AtxtComplemento;
+    private javax.swing.JComboBox<String> CboEndereco;
+    private javax.swing.JDesktopPane DesktopMain;
+    private javax.swing.JTextField btnNomeCli;
+    private datechooser.beans.DateChooserCombo dateChooserCombo1;
+    private javax.swing.JButton jButton1;
+    private javax.swing.JButton jButton2;
+    private javax.swing.JComboBox<String> jComboBox2;
+    private javax.swing.JFormattedTextField jFormattedTextField1;
+    private javax.swing.JLabel lblCEP;
+    private javax.swing.JLabel lblCPF;
+    private javax.swing.JLabel lblCidade;
+    private javax.swing.JLabel lblCodFidelidade;
+    private javax.swing.JLabel lblComplemento;
+    private javax.swing.JLabel lblDataNacimento;
+    private javax.swing.JLabel lblEmail;
+    private javax.swing.JLabel lblEndereco;
+    private javax.swing.JLabel lblEstado;
+    private javax.swing.JLabel lblFone;
+    private javax.swing.JLabel lblInfComplement;
+    private javax.swing.JLabel lblNome;
+    private javax.swing.JLabel lblNumero;
+    private javax.swing.JFormattedTextField txtCEP;
+    private javax.swing.JFormattedTextField txtCPF;
+    private javax.swing.JTextField txtCidade;
+    private javax.swing.JTextField txtCodFidelidade;
+    private javax.swing.JTextField txtComplemento;
+    private javax.swing.JTextField txtEmail;
+    private javax.swing.JTextField txtEndereco;
+    private javax.swing.JFormattedTextField txtFone;
+    // End of variables declaration//GEN-END:variables
+}
