--- src/gabenricks_h/c/Gabenricks_HC.java
@@ -1,21 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package gabenricks_h.c;
-
-/**
- *
- * @author HOLANDAS
- */
-public class Gabenricks_HC {
-
-    /**
-     * @param args the command line arguments
-     */
-    public static void main(String[] args) {
-        // TODO code application logic here
-    }
-    
-}
